python_core/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
passlib[bcrypt]
pyjwt
boto3